[Rename] : Community -> community  폴더 이름 변경
