+++ client/src/ccc/common/Footer.tsx
