+++ client/src/ccc/community/ReviewCarousel.tsx
