+++ client/src/ccc/community/SearchPlace.tsx
