+++ client/src/ccc/mainpage/Banner.tsx
