+++ client/src/ccc/mainpage/Carousel.tsx
