+++ client/src/ccc/mainpage/CarouselNearbyPlace.tsx
