+++ client/src/pages/Community/PostUpload.tsx
@@ -15,8 +15,8 @@ import colorSyntax from '@toast-ui/editor-plugin-color-syntax';
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import { Editor } from '@toast-ui/react-editor';
 
-import SearchPlace from '../../Components/community/SearchPlace';
-import SubjectDropdown from '../../Components/community/SubjectDropdown';
+import SearchPlace from '../../ccc/community/SearchPlace';
+import SubjectDropdown from '../../ccc/community/SubjectDropdown';
 import { Api } from '../../apis/customAPI';
 import useAxios from '../../hooks/useAxios';
 import { Iuser, UPDATE } from '../../reducers/userInfoReducer';
